+++ homely/general.py
@@ -1,17 +1,23 @@
-import io
 import os
+import sys
 from contextlib import contextmanager
-from importlib.machinery import SourceFileLoader
 
 from homely._engine2 import Engine, Helper, getengine, getrepoinfo
 from homely._ui import entersection, warn
 # allow importing from outside
 from homely._utils import haveexecutable  # noqa
-from homely._utils import _homepath2real, _repopath2real
+from homely._utils import _homepath2real, _loadmodule, _repopath2real
 # TODO: remove these deprecated aliases which I'm still using in my homely
-# repos
+# repos. Note that the cleaners will need some sort of special handling in
+# cleanerfromdict() if ever we want to remove these imports
 from homely.files import (WHERE_ANY, WHERE_BOT, WHERE_END, WHERE_TOP,  # noqa
-                          blockinfile, download, lineinfile, mkdir, symlink)
+                          CleanBlockInFile, CleanLineInFile, blockinfile,
+                          download, lineinfile, mkdir, symlink)
+
+if sys.version_info[0] < 3:
+    from StringIO import StringIO
+else:
+    from io import StringIO
 
 
 def run(updatehelper):
@@ -30,13 +36,13 @@ def include(pyscript):
     global _include_num
     _include_num += 1
 
-    source = SourceFileLoader('__imported_by_homely_{}'.format(_include_num),
-                              path)
+    name = '__imported_by_homely_{}'.format(_include_num)
     try:
         with entersection("/" + pyscript):
-            source.load_module()
+            _loadmodule(name, path)
     except Exception as err:
-        warn("Error while including {}: {}".format(pyscript, str(err)))
+        import traceback
+        warn("Error while including {}: {}".format(pyscript, traceback.format_exc()))
 
 
 def section(func):
@@ -54,7 +60,7 @@ def section(func):
 def writefile(filename):
     stream = None
     try:
-        stream = io.StringIO('foo')
+        stream = StringIO()
         yield stream
         stream.seek(0)
         getengine().run(WriteFile(_homepath2real(filename), stream.read()))
@@ -80,16 +86,15 @@ class WriteFile(Helper):
     def isdone(self):
         if os.path.islink(self._filename):
             return False
-        try:
+        if os.path.exists(self._filename):
             with open(self._filename, 'r') as f:
                 if f.read() == self._contents:
                     return True
-        except FileNotFoundError:
-            pass
         return False
 
     def makechanges(self):
-        assert not os.path.islink(self._filename)
+        if os.path.islink(self._filename):
+            raise Exception("Path is already a symlink")
         with open(self._filename, 'w') as f:
             f.write(self._contents)