+++ homely/ui.py
@@ -1,5 +1,4 @@
 # export these items:
-from homely._ui import (  # noqa
-    head, note, warn, yesno, allowpull, allowinteractive, system,
-)
-from homely._errors import InputError  # noqa
+from homely._errors import InputError  # noqa: F401
+from homely._ui import (allowinteractive, allowpull, head, note,  # noqa: F401
+                        system, warn, yesno)